% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association.R
\name{clinical_correlation}
\alias{clinical_correlation}
\title{Correlation test or Kruskal Wallis test (numeric or categorical feature)}
\usage{
clinical_correlation(individual_scores, component, clin_col)
}
\arguments{
\item{individual_scores}{PC1 and PC2 scores for individuals}

\item{component}{principal component to look at (1,2)}

\item{clin_col}{clinical feature for individuals (e.g. age)}
}
\value{
pvalue
}
\description{
Correlation test or Kruskal Wallis test (numeric or categorical feature)
}
