% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_random.R
\name{create_gene_set}
\alias{create_gene_set}
\title{Creates gene set from a universe of genes}
\usage{
create_gene_set(universe, psize, n_perm = 1000)
}
\arguments{
\item{universe}{Gene universe to sample random gene set from}

\item{psize}{Number of genes in a gene set}

\item{n_perm}{Number of permutations to create a random gene set
(default: 1000)}
}
\value{
Random gene set
}
\description{
This function creates random gene sets
}
