% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathways.R
\name{filter_pathways}
\alias{filter_pathways}
\title{Filter a pathway list}
\usage{
filter_pathways(pathways_list, edges)
}
\arguments{
\item{pathways_list}{list of pathways}

\item{edges}{Table, containing information on "reg" and "tar"}
}
\value{
A list of filtered pathways
}
\description{
This function filters a list of pathways to include only genes in pathways
present in networks
}
