% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathways.R
\name{filter_pathways_size}
\alias{filter_pathways_size}
\title{Filter a pathway list based on a number of genes in a pathway}
\usage{
filter_pathways_size(pathways_list, minSize = 5, maxSize = 150)
}
\arguments{
\item{pathways_list}{list of pathways}

\item{minSize}{Minimum size for number of genes in a pathway (default: 5)}

\item{maxSize}{Maximum size for a number of genes in a pathway (default: 150)}
}
\value{
A list of filtered pathways
}
\description{
This function filters a list of pathways based on specified minimum and
maximum size for number of genes in a pathway
}
