% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pt.R
\name{filter_pt}
\alias{filter_pt}
\title{Filter a pathway file}
\usage{
filter_pt(pt.list, net, length.thres)
}
\arguments{
\item{pt.list}{List of pathways}

\item{net}{A numeric matrix or a data frame, representing network}

\item{length.thres}{Threshold to filter pathways, based on their length}
}
\value{
A list of filtered pathways
}
\description{
This function filters a list of pathways.
}
