% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_scores.R
\name{get_features_scores}
\alias{get_features_scores}
\title{Extract feature scores on the first two principal components}
\usage{
get_features_scores(data, scale_data = TRUE, center_data = TRUE)
}
\arguments{
\item{data}{Numeric matrix with samples in columns, and features in rows}

\item{scale_data}{Whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{center_data}{Logical, whether to center the data (TRUE) or not (FALSE),
Default is TRUE.}
}
\value{
features scores on the PC1 and PC2
}
\description{
This function extracts feature scores on the first two principal components
}
