% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association.R
\name{get_ind_scores}
\alias{get_ind_scores}
\title{Extract individual scores from the PCA results}
\usage{
get_ind_scores(data, scale_data = TRUE)
}
\arguments{
\item{data}{Numeric matrix with samples in columns, and features in rows}
}
\value{
individual scores for PC1 and PC2
}
\description{
Extract individual scores from the PCA results
}
