% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association.R
\name{get_pathway_ind_scores}
\alias{get_pathway_ind_scores}
\title{Extract individual scores from PCA results for a list of pathways}
\usage{
get_pathway_ind_scores(pathways_list, reg_net, edges, scale_data = TRUE)
}
\arguments{
\item{pathways_list}{list of pathways}

\item{reg_net}{Numeric matrix with samples in columns, features in rows}

\item{edges}{Table, containing information on "reg" and "tar"}

\item{data}{Numeric matrix with samples in columns, and features in rows}

\item{ncores}{A number of cores to use (default: 1)}
}
\value{
individual scores for each pathway
}
\description{
Extract individual scores from PCA results for a list of pathways
}
