% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gmt.R
\name{load_gmt}
\alias{load_gmt}
\title{Load a .gmt file}
\usage{
load_gmt(gmt_file)
}
\arguments{
\item{gmt_file}{Path to the input .gmt file. Gmt file is downloaded from http://www.gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\value{
A list of pathways from the .gmt file
}
\description{
This function loads a .gmt file.
}
