% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_pathway.R
\name{pca_pathway}
\alias{pca_pathway}
\title{Run PCA analysis for a list of pathways performed on network edges}
\usage{
pca_pathway(
  pathways_list,
  reg_net,
  edges,
  ncores = 1,
  scale_data = TRUE,
  center_data = TRUE
)
}
\arguments{
\item{pathways_list}{list of pathways}

\item{reg_net}{Numeric matrix with samples in columns, features in rows}

\item{edges}{Table, containing information on "reg" and "tar"}

\item{ncores}{A number of cores to use (default: 1)}

\item{scale_data}{Logical, whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{center_data}{Logical, whether to center the data (TRUE) or not (FALSE),
Default is TRUE.}
}
\value{
Dataframe with pca results for pathways in a pathway file
}
\description{
This function runs PCA analysis for a list of pathways
}
