% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_pt.R
\name{pca_pt}
\alias{pca_pt}
\title{Run PCA analysis for a list of pathways on network edges}
\usage{
pca_pt(pt.filt, net, ncores)
}
\arguments{
\item{pt.filt}{A list of filtered pathways}

\item{net}{A numeric matrix or a data frame, representing network}

\item{ncores}{A number of cores to use}
}
\value{
Dataframe with pca results for pathways in pathway file
}
\description{
This function runs PCA analysis for a list of pathways
}
