% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_random.R
\name{pca_random}
\alias{pca_random}
\title{Creates random gene sets and runs PCA analysis on network for a set of random
gene sets}
\usage{
pca_random(
  reg_net,
  edges,
  results_pca_pathways,
  pathways_list,
  n_perm = 1000,
  ncores = 1,
  scale_data = TRUE
)
}
\arguments{
\item{reg_net}{Table of network with samples in columns, features in rows}

\item{edges}{Table, containing information on "reg" and "tar" of reg_net}

\item{results_pca_pathways}{Output result table of pca_pathway function}

\item{pathways_list}{A list of pathways}

\item{n_perm}{Number of permutations to create a random gene set
(default: 1000)}

\item{ncores}{A number of cores to use (default: 1)}
}
\value{
Dataframe with pca results for random gene sets
}
\description{
This function creates random gene sets and runs PCA analysis on a set of
random gene sets
}
