% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association.R
\name{plot_clinical_association}
\alias{plot_clinical_association}
\title{Plot PCA colored individuals based on the clinical feature}
\usage{
plot_clinical_association(scores, clin_col)
}
\arguments{
\item{clin_col}{clinical feature for individuals (e.g. age)}

\item{individual_scores}{PC1 and PC2 scores for individuals}
}
\value{
PCA plot with individuals colored according to clinical feature
}
\description{
Plot PCA colored individuals based on the clinical feature
}
