% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PORCUPINE.R
\name{PORCUPINE}
\alias{PORCUPINE}
\title{Determine statistical significance of PCA scores of pathways}
\usage{
PORCUPINE(res_pca_pathways, res_pca_rndm, test = t.test)
}
\arguments{
\item{res_pca_pathways}{Output result table of pca_pathway function}

\item{res_pca_random}{Output result table of pca_random function}
}
\value{
Table of statistics (pvalue and effect size) for each pathway
}
\description{
This function compares results of PCA analyses for each pathway versus set of permutated gene sets. Calculates p-value and effect size for each pathway.
}
