% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_visualize_clusters.R
\name{select_number_clusters}
\alias{select_number_clusters}
\title{Visualize and select an optimal number of clusters of individuals based
on the network information for a pathway of interest}
\usage{
select_number_clusters(
  pathway_of_interest,
  reg_net,
  edges,
  scale_data = TRUE,
  center_data = TRUE,
  kmax = 8
)
}
\arguments{
\item{pathway_of_interest}{List with genes in a pathway of interest}

\item{reg_net}{Numeric matrix with samples in columns, features in rows}

\item{edges}{Data frame containing information on "reg" and "tar"}

\item{scale_data}{Whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{center_data}{Whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{k_max}{Integer specifying the maximum nuber of clusters to consider}
}
\value{
ggplot object
}
\description{
This function fertemines and visualizes the optimal number of clusters using
kmeans clustering
}
