% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_visualize_clusters.R
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize clustering results}
\usage{
visualize_clusters(
  pathway_of_interest,
  reg_net,
  edges,
  scale_data = TRUE,
  center_data = TRUE,
  number_of_clusters
)
}
\arguments{
\item{pathway_of_interest}{List with genes in a pathway of interest}

\item{reg_net}{Numeric matrix with samples in columns, features in rows}

\item{edges}{Data frame containing information on "reg" and "tar"}

\item{scale_data}{Whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{center_data}{Whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{number_of_clusters}{Integer, specifying the number of clusters}
}
\value{
ggplot object
}
\description{
Visuzalition of clustering of patients into specified number of clusters
}
