% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_scores.R
\name{clinical_correlation}
\alias{clinical_correlation}
\title{Significance of association of clinical feature with patient heterogeneity
scores (PC1 or PC2 scores)}
\usage{
clinical_correlation(individual_scores, component, clin_feature)
}
\arguments{
\item{individual_scores}{PC1 and PC2 heterogeneity scores for individuals}

\item{component}{principal component to look at (1,2)}

\item{clin_feature}{clinical feature for individuals (e.g. age)}
}
\value{
pvalue
}
\description{
This function performs a correlation test or Kruskal-Wallis test (for numeric
or categorical variables, respectively) between patient heterogeneity scores
on PC1 or PC2
}
