% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gmt.R
\name{load_gmt}
\alias{load_gmt}
\title{Load a gmt file}
\usage{
load_gmt(gmt.file)
}
\arguments{
\item{gmt.file}{Path to the input gmt file}
}
\value{
A list of the gmt.file
}
\description{
This function loads a gmt file.
}
