% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_pathway.R
\name{pca_pathway}
\alias{pca_pathway}
\title{Run PCA analysis for a list of pathways on network edges}
\usage{
pca_pathway(pathways_list, RegNet, edges, ncores)
}
\arguments{
\item{pathways_list}{list of pathways}

\item{RegNet}{Numeric matrix with samples in columns}

\item{edges}{Table, containing information on "reg" and "tar"}

\item{ncores}{A number of cores to use}
}
\value{
Dataframe with pca results for pathways in a pathway file
}
\description{
This function runs PCA analysis for a list of pathways
}
