% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_random.R
\name{pca_random}
\alias{pca_random}
\title{Creates random gene sets and runs PCA analysis on a set of random gene sets}
\usage{
pca_random(net, res.ptws, pt.filt, n.perm, ncores)
}
\arguments{
\item{net}{A numeric matrix or a data frame, representing network}

\item{res.ptws}{Output result table of pca_gmt function}

\item{pt.filt}{A list of filtered pathways}

\item{n.perm}{Number of permutations to create random gene set}

\item{ncores}{A number of cores to use}
}
\value{
Dataframe with pca results for random gene sets
}
\description{
This function creates random gene sets and runs PCA analysis on a set of random gene sets
}
