% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_scores.R
\name{plot_clinical_association}
\alias{plot_clinical_association}
\title{PCA plot of individuals colored according to a clinical feature}
\usage{
plot_clinical_association(individual_scores, clin_col)
}
\arguments{
\item{individual_scores}{individual heterogeneity scores on PC1 and PC2}

\item{clin_feature}{clinical feature for individuals (e.g. age)}
}
\value{
PCA plot
}
\description{
This function makes a PCA plot of individuals colored according to a
clinical feature
}
