% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PORCUPINE.R
\name{PORCUPINE}
\alias{PORCUPINE}
\title{Compares PCA results for each pathway versus PCA results of a set set of random gene sets of the same size, calculates p-value and effect size for each pathway}
\usage{
PORCUPINE(res.ptws, res.random)
}
\arguments{
\item{res.ptws}{Output result table of pca_gmt function}

\item{res.random}{Output result table of pca_random function}
}
\description{
This function compares results of PCA analyses for each pathway versus set of permutated pathways of the same size, calculates p-value and effect size for each pathway
}
