% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA analysis on the data}
\usage{
run_pca(data, scale_data = TRUE, center_data = TRUE)
}
\arguments{
\item{data}{Numeric matrix with samples in columns, and features in rows}

\item{scale_data}{Logical, whether to scale the data (TRUE) or not (FALSE),
Default is TRUE.}

\item{center_data}{Logical, whether to center the data (TRUE) or not (FALSE),
Default is TRUE.}
}
\value{
Dataframe with pca results
}
\description{
This function perform PCA analysis on the data
}
