% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA analysis on network data}
\usage{
run_pca(data)
}
\arguments{
\item{data}{Numeric matrix with samples in columns}
}
\value{
Dataframe with pca results
}
\description{
This function runs PCA analysis
}
